+++ seqcluster/libs/parse.py
@@ -11,6 +11,7 @@ def parse_cl(in_args):
                 "predict": add_subparser_predict,
                 "explore": add_subparser_explore,
                 "collapse": add_subparser_collapse,
+                "simulator": add_subparser_simulator,
                 "stats": add_subparser_stats}
     parser = argparse.ArgumentParser(description="small RNA analysis")
     sub_cmd = None
@@ -262,11 +263,12 @@ def add_subparser_collapse(subparsers):
 
 def add_subparser_simulator(subparsers):
     parser = subparsers.add_parser("simulator", help="simulate small RNA  from bed file")
-    parser.add_argument("--bed", dest="bed", required=1,
+    parser.add_argument("--bed",
                         help="bed file with position of precursors <=200 nt")
+    parser.add_argument("--fasta", help = "fasta with precursors.")
     parser.add_argument("--out", dest="out", required=1,
                         help="dir of output files")
-    parser.add_argument("-r", "--reference", dest="ref", required=1,
+    parser.add_argument("-r", "--reference", dest="ref",
                         help="reference fasta file with index"),
     parser = _add_debug_option(parser)
     return parser