+++ HISTORY.md
@@ -1,5 +1,7 @@
 - current
 
+* first version of simulator of 33 nt long sRNA from fasta
+* add Rmd template for bcbio pipeline
 * add peak detection to cluster, as a try to get the mature sequence
 
 - 1.1.8